requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.31.1
fpdf2==2.5.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
kiwisolver==1.4.0
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pick==1.2.0
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
shellingham==1.4.0
six==1.16.0
tabulate==0.8.9
tomli==2.0.1
tornado==6.1
typer==0.4.0
typing-extensions==4.1.1
urllib3==1.26.8
